Moved hard-coded values to settings; remove active_window calls
The settings now can specify the regex for compile errors. Added a function to getHostType from the ACCUTERM file. This can be expanded to load the DBMS type from the settings file (at this time nothing would use this). Thanks to @Thom1729 for suggesting the removal of active_window() calls, which will prevent bugs caused by window switching.