Enable compile command to accept a list
The compile_command setting now can accept a list to run multiple commands in the form:
[command] [mv file] [mv item]
This enables jBASE's need to compile and then catalog. This feature needs to be expanded to enable specifying where the mv file and item are in the compile command. This would enable commands to be run without automatically appending the file and item to them. The setting should also handle various commands based on the DBMS type like the result regex does.