Enabled locking based on view setting
Added a view setting to track the lock state of the item on the MV server. If the item was opened without locking then locking will not be preformed on subsequent opens (unless manually set). This setting is preserved when Sublime is closed so an enhancement will need to be made to check locks when it's reopened.

Fixes #5