Moved the locking & sync check
Altered the view event listener so the locking update and sync check only happen once the view is activated after loading. Bringing up the file switcher panel was causing the load event to fire which would then steal the view focus when locking was updated. Also the sync check would pop a dialog box if the file was out of sync before the item was selected from the panel. 

Now the view event listener sets a flag when the load event is fired that is required by the activate event to check sync & update locks. The activate event un-sets the flag so the functions will only be run once after loading.